+++ dbt-ml/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## v0.8.0 - 2026-08-11
 
 ### Coalesce small incremental flushes into one publication (issue #293)